pyproject.toml
[tool.poetry]
name = "ibkr-portfolio-bot"
version = "0.1.0"
description = "Production-grade IBKR fractional swing trading system"
authors = ["Your Name <you@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
ib-insync = "^0.9.86"
pandas = "^2.1.0"
numpy = "^1.26.0"
pyarrow = "^14.0.0"
matplotlib = "^3.8.0"
click = "^8.1.7"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pytz = "^2023.3"
pandas-market-calendars = "^4.3.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.21.1"
ruff = "^0.1.6"
mypy = "^1.7.0"
types-pytz = "^2023.3.0.10"

[tool.poetry.scripts]
bot = "src.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "ib_insync.*",
    "pandas_market_calendars.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=json",
    "--cov-fail-under=85",
    "--maxfail=10",  # Stop after 10 failures
]
asyncio_mode = "auto"
markers = [
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
]
